% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year}
\alias{check_year}
\title{Checks if year is valid}
\usage{
check_year(year = NULL, min = NULL)
}
\arguments{
\item{year}{Year.}

\item{min}{Numeric for minimum year available.}
}
\description{
Checks if year is valid
}
\keyword{internal}
