% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lovb_player_data}
\alias{lovb_player_data}
\title{League One Volleyball Individual Player Statistics}
\format{
A data frame with 94 rows and 27 variables:
\describe{
\item{year}{Season year}
\item{team}{Player's team name}
\item{number}{Player jersey number}
\item{player}{Player name}
\item{points}{Total Points}
\item{matches_started}{Matches Started}
\item{sets_started}{Sets Started}
\item{sets_played}{Sets Played}
\item{hitting_efficiency}{Attack Efficiency percentage}
\item{kill_percentage}{Kill percentage}
\item{kills}{Kills - successful attacks}
\item{attack_errors}{Attack Errors}
\item{attacks_blocked}{Attacks blocked by opponent}
\item{attack_attempts}{Total attack attempts}
\item{in_system_percentage}{In-system percentage}
\item{reception_errors}{Reception errors}
\item{reception_attempts}{Reception attempts}
\item{service_aces}{Service Aces - serves that result directly in a point}
\item{service_errors}{Service errors}
\item{opponent_in_system_percentage}{Opponent in-system percentage}
\item{service_attempts}{Service attempts}
\item{blocks}{Blocks - successful defensive plays at the net}
\item{block_touch_percentage}{Block touch percentage}
\item{digs}{Digs - successful defensive plays}
\item{dig_percentage}{Dig percentage}
\item{assists}{Assists - sets that lead directly to a kill}
\item{setting_efficiency}{Setting efficiency percentage}
}
}
\source{
\url{https://lovb.com}
}
\usage{
lovb_player_data
}
\description{
A dataset containing individual player season statistics for League One Volleyball (LOVB) players
across the league for the 2025 season.
}
\examples{
head(lovb_player_data)

}
\seealso{
Other datasets:
\code{\link{au_player_data}},
\code{\link{lovb_team_data}},
\code{\link{lovb_teams}},
\code{\link{mlv_player_data}},
\code{\link{mlv_team_data}},
\code{\link{mlv_teams}}
}
\concept{datasets}
\keyword{datasets}
