% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lovb_team_data}
\alias{lovb_team_data}
\title{League One Volleyball Team Match Statistics}
\format{
A data frame with 96 rows and 24 variables:
\describe{
\item{year}{Season year}
\item{team}{Team name}
\item{opponent}{Opposing team name}
\item{date}{Match date}
\item{points}{Points scored}
\item{hitting_efficiency}{Attack efficiency percentage}
\item{kill_percentage}{Kill percentage}
\item{kills}{Kills - successful attacks that result in a point}
\item{attack_errors}{Attack Errors - unsuccessful attacks}
\item{attacks_blocked}{Attacks blocked by opponent}
\item{attack_attempts}{Total attack attempts}
\item{in_system_percentage}{In-system percentage}
\item{reception_errors}{Reception errors}
\item{reception_attempts}{Reception attempts}
\item{service_aces}{Service Aces - serves that result directly in a point}
\item{service_errors}{Service errors}
\item{opponent_in_system_percentage}{Opponent in-system percentage}
\item{service_attempts}{Service attempts}
\item{blocks}{Blocks - successful defensive plays at the net}
\item{block_touch_percentage}{Block touch percentage}
\item{digs}{Digs - successful defensive plays}
\item{dig_percentage}{Dig percentage}
\item{assists}{Assists - sets that lead directly to a kill}
\item{setting_efficiency}{Setting efficiency percentage}
}
}
\source{
\url{https://lovb.com}
}
\usage{
lovb_team_data
}
\description{
A dataset containing team match-by-match statistics for League One Volleyball (LOVB) teams
for the 2025 season.
}
\examples{
head(lovb_team_data)

}
\seealso{
Other datasets:
\code{\link{au_player_data}},
\code{\link{lovb_player_data}},
\code{\link{lovb_teams}},
\code{\link{mlv_player_data}},
\code{\link{mlv_team_data}},
\code{\link{mlv_teams}}
}
\concept{datasets}
\keyword{datasets}
