% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lovb_teams}
\alias{lovb_teams}
\title{League One Volleyball Teams}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{name}{Team city name}
\item{slug}{URL slug used in LOVB website URLs}
\item{first_year}{Year that team started in LOVB}
}
}
\source{
\url{https://www.lovb.com}
}
\usage{
lovb_teams
}
\description{
A dataset containing information about League One Volleyball (LOVB) teams,
including team names, URL slugs, and first year used for web scraping.
}
\details{
This dataset is used internally by the package functions to map
user-provided team names to the appropriate URL slugs needed for web scraping
the LOVB website.
}
\examples{
lovb_teams

}
\seealso{
Other datasets:
\code{\link{au_player_data}},
\code{\link{lovb_player_data}},
\code{\link{lovb_team_data}},
\code{\link{mlv_player_data}},
\code{\link{mlv_team_data}},
\code{\link{mlv_teams}}
}
\concept{datasets}
\keyword{datasets}
