% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvf_team_data}
\alias{pvf_team_data}
\title{Pro Volleyball Federation Team Match Statistics}
\format{
A data frame with 392 rows and 15 variables:
\describe{
\item{year}{Season year}
\item{team}{Team name}
\item{date}{Match date}
\item{opponent}{Opposing team name}
\item{result}{Win/Loss result}
\item{kills}{Kills - successful attacks that result in a point}
\item{assists}{Assists - sets that lead directly to a kill}
\item{service_aces}{Service Aces - serves that result directly in a point}
\item{blocks}{Blocks - successful defensive plays at the net}
\item{out}{Definition pending}
\item{attack_attempts}{Total attack attempts}
\item{hitting_efficiency}{Attack efficiency percentage}
\item{digs}{Digs - successful defensive plays}
\item{digs_per_set}{Digs per set}
\item{sets_played}{Sets played in the match}
}
}
\source{
\url{https://provolleyball.com}
}
\usage{
pvf_team_data
}
\description{
A dataset containing team match-by-match statistics for Pro Volleyball Federation (PVF) teams
for the 2024-2025 seasons.
}
\examples{
head(pvf_team_data)

}
\seealso{
Other datasets: 
\code{\link{lovb_player_data}},
\code{\link{lovb_team_data}},
\code{\link{lovb_teams}},
\code{\link{pvf_player_data}},
\code{\link{pvf_teams}}
}
\concept{datasets}
\keyword{datasets}
