% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{au_player_data}
\alias{au_player_data}
\title{Athletes Unlimited Individual Player Statistics}
\format{
A data frame with 220 rows and 25 variables:
\describe{
\item{year}{Season year}
\item{rank}{Player rank}
\item{player}{Player name}
\item{points}{Total points achieived}
\item{sets_played}{Sets played}
\item{kills}{Kills - successful attacks}
\item{kills_per_set}{Kills per set}
\item{attack_errors}{Attack Errors}
\item{attack_attempts}{Total attack attempts}
\item{kill_percentage}{Kill percentage}
\item{assists}{Assists - sets that lead directly to a kill}
\item{assists_per_set}{Assists per set}
\item{setting_errors}{Setting errors}
\item{service_aces}{Service Aces - serves that result directly in a point}
\item{service_errors}{Service errors}
\item{service_aces_per_set}{Service aces per set}
\item{total_reception_attempts}{Total reception attempts}
\item{reception_errors}{Reception errors}
\item{positive_reception_percentage}{Percentage of reception attempts that are either positive or perfect}
\item{digs}{Digs - successful defensive plays}
\item{digs_per_set}{Digs per set}
\item{blocks}{Blocks - successful defensive plays at the net}
\item{blocks_per_set}{Blocks per set}
\item{block_assists}{Block assists - two- or three-person blocks}
\item{good_receptions}{Number of positive receptions}
}
}
\source{
\url{https://auprosports.com/volleyball}
}
\usage{
au_player_data
}
\description{
A dataset containing individual player season statistics for Athletes Unlimited (UA) players
for the 2021-2025 seasons.
}
\examples{
head(au_player_data)

}
\seealso{
Other datasets: 
\code{\link{lovb_player_data}},
\code{\link{lovb_team_data}},
\code{\link{lovb_teams}},
\code{\link{mlv_player_data}},
\code{\link{mlv_team_data}},
\code{\link{mlv_teams}}
}
\concept{datasets}
\keyword{datasets}
