% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Extract Data from US Women's Pro Volleyball Websites}
\usage{
get_stats(league = NULL, team = NULL, year = NULL, level = NULL)
}
\arguments{
\item{league}{A character string specifying which league to retrieve. Must be
"LOVB" or "PVF".}

\item{team}{A character string specifying the team name or city. Accepts
city names (e.g., "Omaha", "Atlanta") or variations like "Indianapolis"
(converted to "Indy") and "Las Vegas" (converted to "Vegas"). Must match
one of the valid LOVB or PVF team cities.}

\item{year}{A numeric value specifying the year for which to retrieve
statistics. Must be 2024 or later, corresponding to available LOVB or PVF seasons.}

\item{level}{A character string specifying whether to extract the "team" or individual
"player" statistics.}
}
\value{
A tibble containing team match-by-match or player season statistics.
}
\description{
This function scrapes team match-by-match and player season statistics for a specified
League One Volleyball (LOVB) or Pro Volleyball Federation (PVF) team and
year from the official website.
}
\note{
This function requires:
\itemize{
\item The \if{html}{\out{<a href="https://ashbythorpe.github.io/selenider/">}}\if{html}{\out{<code>}}\{selenider\}\if{html}{\out{</code>}}\if{html}{\out{</a>}} R package
for web automation
\item \href{https://www.google.com/chrome/}{Google Chrome} browser installed on the system
\item An active internet connection
}

The function uses a headless Chrome browser session, so no browser window
will be visible during execution. Browser sessions are automatically closed
after data extraction.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get Omaha Supernovas 2024 match statistics
get_stats(league = "PVF", team = "Omaha", year = 2024, level = "team")
get_stats(league = "PVF", team = "Omaha", year = 2024, level = "player")
get_stats(league = "LOVB", team = "Austin", year = 2025, level = "team")
get_stats(league = "LOVB", team = "Austin", year = 2025, level = "player")
\dontshow{\}) # examplesIf}
}
\seealso{
Other statistics functions: 
\code{\link{group_stats}()},
\code{\link{lovb_stats}()},
\code{\link{pvf_stats}()}
}
\concept{statistics functions}
