% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlv_player_data}
\alias{mlv_player_data}
\title{Major League Volleyball Individual Player Statistics}
\format{
A data frame with 264 rows and 23 variables:
\describe{
\item{year}{Season year}
\item{team}{Player's team name}
\item{number}{Player jersey number}
\item{player}{Player name}
\item{sets_played}{Sets Played}
\item{matches_played}{Matches Played}
\item{points}{Total Points}
\item{points_per_set}{Points per Set}
\item{kills}{Kills - successful attacks}
\item{kills_per_set}{Kills per Set}
\item{attack_errors}{Attack Errors}
\item{attack_attempts}{Total Attacks}
\item{hitting_efficiency}{Attack Efficiency percentage}
\item{assists}{Assists - sets that lead directly to a kill}
\item{assists_per_set}{Assists per Set}
\item{service_aces}{Service Aces - serves that result directly in a point}
\item{service_aces_per_set}{Service Aces per Set}
\item{serive_errors}{Service Errors}
\item{service_errors_per_set}{Service Errors per Set}
\item{digs}{Digs - successful defensive plays}
\item{digs_per_set}{Digs per Set}
\item{blocks}{Blocks - successful defensive plays at the net}
\item{blocks_per_set}{Blocks per Set}
}
}
\source{
\url{https://provolleyball.com}
}
\usage{
mlv_player_data
}
\description{
A dataset containing individual player season statistics for Major League Volleyball (MLV) players
across the league for the 2024-2025 seasons.
}
\examples{
head(mlv_player_data)

}
\seealso{
Other datasets: 
\code{\link{au_player_data}},
\code{\link{lovb_player_data}},
\code{\link{lovb_team_data}},
\code{\link{lovb_teams}},
\code{\link{mlv_team_data}},
\code{\link{mlv_teams}}
}
\concept{datasets}
\keyword{datasets}
