% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provolleyballr-package.R
\docType{package}
\name{provolleyballr-package}
\alias{provolleyballr}
\alias{provolleyballr-package}
\title{provolleyballr: Extract Data from US Women's Professional Volleyball Websites}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for scraping match statistics and player data from the Athletes Unlimited (UA) website \url{https://auprosports.com/volleyball}, the League One Volleyball website \url{https://lovb.com}, and the Major League (MLV) website \url{https://provolleyball.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeffreyRStevens/provolleyballr}
  \item Report bugs at \url{https://github.com/JeffreyRStevens/provolleyballr/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@protonmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [copyright holder]

}
\keyword{internal}
